+++ config/settings/base.py
@@ -11,9 +11,9 @@ from django.contrib.messages import constants as messages
 # 3rd Party Libraries
 import environ
 
-__version__ = "3.1.2"
+__version__ = "3.1.3"
 VERSION = __version__
-RELEASE_DATE = "31 October 2022"
+RELEASE_DATE = "03 November 2022"
 
 ROOT_DIR = Path(__file__).resolve(strict=True).parent.parent.parent
 APPS_DIR = ROOT_DIR / "ghostwriter"