+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.3] - 3 November 2022
+
+### Added
+
+* Added a new configuration section with the ability to set a default timezone used for forms
+* Added a new option to Global Report Configuration to enable configuring default download filenames for report files
+
 ## [3.1.2] - 31 October 2022
 
 ### Fixed